Jabalizer Documentation

Contents

Jabalizer Documentation

Contents

Functions

Index

Functions

Main.Jabalizer.AddQubit — Function

Add a qubit to a stabilizer with given Pauli operator and phase.

Add a qubit to a stabilizer with identity operator.

Main.Jabalizer.AddQubits — Function

Add multiple qubits to a Stabilizer.

Main.Jabalizer.AddBell — Function

AddBell(state, labelA, labelB)

Add Bell state $(|00\rangle+|11\rangle)/\sqrt{2}$ to a State.

AddBell(state)

Add Bell state $(|00\rangle+|11\rangle)/\sqrt{2}$ to a State.

Main.Jabalizer.AddGHZ — Function

AddGHZ(state, size, labels)

Add a GHZ state to a State.

AddGHZ(state, size)

Add a GHZ state to a State.

AddGHZ(state)

Add a 3-qubit GHZ state to a State.

Main.Jabalizer.AddGraph — Function

AddGraph(state, graph, labels)

Add a graph state to a State.

AddGraph(state, graph)

Add a graph state to a State.

Main.Jabalizer.CNOT — Function

CNOT(Stabilizer, control, target)

Apply CNOT gate to a Stabilizer.

CNOT(State, control, target)

Apply CNOT gate to a State.

Main.Jabalizer.CZ — Function

CZ(State, control, target)

Apply CZ gate to a Stabilizer.

CZ(State, control, target)

Apply CZ gate to a State.

Main.Jabalizer.H — Function

Apply H gate to a Stabilizer.

H(State, qubit)

Apply H gate to a State.

Main.Jabalizer.ChannelX — Function

ChannelX(State, qubit, prob)

Apply Pauli-X channel gate to a State.

ChannelX(State, prob)

Apply Pauli-X channel gate to a State.

Main.Jabalizer.ChannelY — Function

ChannelY(State, qubit, prob)

Apply Pauli-Y channel gate to a State.

ChannelY(State, prob)

Apply Pauli-Y channel gate to a State.

Main.Jabalizer.ChannelZ — Function

ChannelZ(State, qubit, prob)

Apply Pauli-Z channel gate to a State.

ChannelZ(State, prob)

Apply Pauli-Z channel gate to a State.

Main.Jabalizer.ChannelPauli — Function

ChannelPauli(State, qubit, probs)

Apply general Pauli channel gate to a State.

ChannelPauli(State, prob)

Apply general Pauli channel gate to a State.

Main.Jabalizer.ChannelDepol — Function

ChannelDepol(State, qubit, prob)

Apply depolarizing channel gate to a State.

ChannelDepol(State, prob)

Apply depolarizing channel gate to a State.

Main.Jabalizer.ChannelLoss — Function

ChannelLoss(State, qubit, prob)

Apply loss channel gate to a State.

ChannelLoss(State, prob)

Apply loss channel gate to a State.

Main.Jabalizer.RowAdd — Function

RowAdd(tableau, source, dest)

Row addition operation for tableaus.

Base.print — Function

print(Stabilizer)

Print a stabilizer to terminal.

print(State)

Print the full stabilizer set of a state to the terminal.

print(GraphState)

Print a GraphState to the terminal.

Base.:* — Function

*(left,right)

Multiplication operator for stabilizers.

Index

Base.:*

Base.print

Main.Jabalizer.AddBell

Main.Jabalizer.AddGHZ

Main.Jabalizer.AddGraph

Main.Jabalizer.AddQubit

Main.Jabalizer.AddQubits

Main.Jabalizer.CNOT

Main.Jabalizer.CZ

Main.Jabalizer.ChannelDepol

Main.Jabalizer.ChannelLoss

Main.Jabalizer.ChannelPauli

Main.Jabalizer.ChannelX

Main.Jabalizer.ChannelY

Main.Jabalizer.ChannelZ

Main.Jabalizer.H

Main.Jabalizer.RowAdd