Jabalizer Documentation

Contents

Jabalizer Documentation

Contents

Types

Functions

Index

Types

Main.Jabalizer.Stabilizer — Type

Stabilizer type

Type for a single stabilizer in the n-qubit Pauli group.

Main.Jabalizer.StabilizerState — Type

Stabilizer state type.

qubits: number of qubits. stabilizers: set of state stabilizers. labels: qubit labels.

Main.Jabalizer.GraphState — Type

Graph state type

Type for a stabilizer state constrained to graph form.

Functions

State preparation

Main.Jabalizer.AddQubit — Function

Add a qubit to a stabilizer with given Pauli operator and phase.

Add a qubit to a stabilizer with identity operator.

Main.Jabalizer.AddQubits — Function

Add multiple qubits to a Stabilizer.

Main.Jabalizer.AddBell — Function

AddBell(state, labelA, labelB)

Add Bell state $(|00\rangle+|11\rangle)/\sqrt{2}$ to a State.

AddBell(state)

Add Bell state $(|00\rangle+|11\rangle)/\sqrt{2}$ to a State.

Main.Jabalizer.AddGHZ — Function

AddGHZ(state, size, labels)

Add a GHZ state to a State.

AddGHZ(state, size)

Add a GHZ state to a State.

AddGHZ(state)

Add a 3-qubit GHZ state to a State.

Main.Jabalizer.AddGraph — Function

AddGraph(state, graph, labels)

Add a graph state to a State.

AddGraph(state, graph)

Add a graph state to a State.

Gates

Main.Jabalizer.Id — Function

Here's an equation:

Id(state, qubit)

Apply I gate to a State on qubit.

Id(state, qubit)

Apply I gate to a State on qubit.

Main.Jabalizer.X — Function

X(stabilizer, qubit)

Apply Pauli-X gate to a stabilizer.

X(state, qubit)

Apply X gate to a State.

Main.Jabalizer.Y — Function

Y(stabilizer, qubit)

Apply Y gate to a stabilizer.

Y(state, qubit)

Apply Y gate to a State.

Main.Jabalizer.Z — Function

Z(stabilizer, qubit)

Apply Z gate to a stabilizer.

Z(state, qubit)

Apply Z gate to a State.

Main.Jabalizer.H — Function

H(stabilizer, qubit)

Apply H gate to a stabilizer.

H(state, qubit)

Apply H gate to a State.

Main.Jabalizer.P — Function

P(stabilizer, qubit)

Apply the $P=\sqrt{Z}$ gate to a stabilizer.

P(state, qubit)

Apply P gate to a state on qubit.

Main.Jabalizer.SWAP — Function

SWAP(stabilizer, qubit1, qubit2)

Apply SWAP gate to a Stabilizer.

SWAP(state, first, second)

Apply SWAP gate to a State.

SWAP(state, first, second)

Apply SWAP gate to a State.

Main.Jabalizer.CNOT — Function

CNOT(Stabilizer, control, target)

Apply CNOT gate to a Stabilizer.

CNOT(state, control, target)

Apply CNOT gate to a State.

Main.Jabalizer.CZ — Function

CZ(state, control, target)

Apply CZ gate to a Stabilizer.

CZ(state, control, target)

Apply CZ gate to a State.

CZ(state, control, target)

Apply CZ gate to a State.

Main.Jabalizer.FusionI — Function

FusionI(state, first, second)

Apply type-I fusion gate to a state.

FusionI(state, first, second)

Apply type-I fusion gate to a state.

Main.Jabalizer.FusionII — Function

FusionII(state, first, second)

Apply type-II fusion gate to a state.

FusionII(state, first, second)

Apply type-II fusion gate to a state.

Channels

Main.Jabalizer.ChannelX — Function

ChannelX(State, qubit, prob)

Apply Pauli-X channel gate to a State.

ChannelX(State, prob)

Apply Pauli-X channel gate to a State.

Main.Jabalizer.ChannelY — Function

ChannelY(State, qubit, prob)

Apply Pauli-Y channel gate to a State.

ChannelY(State, prob)

Apply Pauli-Y channel gate to a State.

Main.Jabalizer.ChannelZ — Function

ChannelZ(State, qubit, prob)

Apply Pauli-Z channel gate to a State.

ChannelZ(State, prob)

Apply Pauli-Z channel gate to a State.

Main.Jabalizer.ChannelPauli — Function

ChannelPauli(State, qubit, probs)

Apply general Pauli channel gate to a State.

ChannelPauli(State, prob)

Apply general Pauli channel gate to a State.

Main.Jabalizer.ChannelDepol — Function

ChannelDepol(State, qubit, prob)

Apply depolarizing channel gate to a State.

ChannelDepol(State, prob)

Apply depolarizing channel gate to a State.

Main.Jabalizer.ChannelLoss — Function

ChannelLoss(State, qubit, prob)

Apply loss channel gate to a State.

ChannelLoss(State, prob)

Apply loss channel gate to a State.

Utilities

Main.Jabalizer.RowAdd — Function

RowAdd(tableau, source, dest)

Row addition operation for tableaus.

Base.print — Function

print(stabilizer)

Print a stabilizer to terminal.

print(state)

Print the full stabilizer set of a state to the terminal.

print(graphState)

Print a GraphState to the terminal.

Base.string — Function

string(stabilizer)

Convert stabilizer to string.

string(state)

Convert state to string.

Base.:* — Function

*(left,right)

Multiplication operator for stabilizers.

Index

Main.Jabalizer.GraphState

Main.Jabalizer.Stabilizer

Main.Jabalizer.StabilizerState

Base.:*

Base.print

Base.string

Main.Jabalizer.AddBell

Main.Jabalizer.AddGHZ

Main.Jabalizer.AddGraph

Main.Jabalizer.AddQubit

Main.Jabalizer.AddQubits

Main.Jabalizer.CNOT

Main.Jabalizer.CZ

Main.Jabalizer.ChannelDepol

Main.Jabalizer.ChannelLoss

Main.Jabalizer.ChannelPauli

Main.Jabalizer.ChannelX

Main.Jabalizer.ChannelY

Main.Jabalizer.ChannelZ

Main.Jabalizer.FusionI

Main.Jabalizer.FusionII

Main.Jabalizer.H

Main.Jabalizer.Id

Main.Jabalizer.P

Main.Jabalizer.RowAdd

Main.Jabalizer.SWAP

Main.Jabalizer.X

Main.Jabalizer.Y

Main.Jabalizer.Z